Move service from data to usecase
