+++ usecase/info_usecase.go
@@ -2,7 +2,7 @@ package usecase
 
 import (
 	"log"
-	"tat_gogogo/data/crawler/service"
+	"tat_gogogo/usecase/service"
 	"tat_gogogo/domain/model"
 	"tat_gogogo/domain/repository"
 )
