+++ usecase/service/result_service.go
